// src/wisdom/shared/enums/artifact-type.enum.ts
// src/wisdom/shared/enums/artifact-type.enum.ts

export enum ArtifactType {

  // Listing
  LISTING_SEARCH_RESULT = "LISTING_SEARCH_RESULT",
  LISTING_SELECTED = "LISTING_SELECTED",

  // Booking
  DATES_SELECTED = "DATES_SELECTED",
  GUEST_COUNT_SELECTED = "GUEST_COUNT_SELECTED",
  CONTACT_SET = "CONTACT_SET",
  SPECIAL_REQUEST_SET = "SPECIAL_REQUEST_SET",
  BOOKING_GET = "BOOKING_GET",
  BOOKING_COMPLETED = "BOOKING_COMPLETED",


  BOOKING_CREATED = "BOOKING_CREATED",
  BOOKING_CONFIRMED = "BOOKING_CONFIRMED",
  BOOKING_CANCELLED = "BOOKING_CANCELLED",

  // AI
  TITLE = "TITLE",
  DESCRIPTION = "DESCRIPTION",
  PRICE = "PRICE",
  SEO = "SEO",
  TIPS = "TIPS",

  REVIEW = "REVIEW",
  BOOKING = "BOOKING",
}